wms-backend/app/requirements.txt
Django==3.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
gunicorn==20.1.0
pytz==2021.1
mysqlclient==2.0.3